+++ python/sglang/srt/layers/moe/ep_moe/pplx_prepare_finalize.py
@@ -1,19 +1,105 @@
 # SPDX-License-Identifier: Apache-2.0
 # Adapted from https://github.com/vllm-project/vllm/blob/f9c069c85e029830094ff9abb926ffbf37b7c7e7/vllm/model_executor/layers/fused_moe/pplx_prepare_finalize.py
 from typing import Optional
-
+import logging
 import pplx_kernels as pplx
+from pplx_kernels.nvshmem import (nvshmem_alloc_empty_unique_id,
+                                          nvshmem_get_unique_id,
+                                  nvshmem_init,
+                                  nvshmem_finalize)
+
 import torch
 
-import sglang.srt.layers.moe.ep_moe.modular_kernel as mk
 from sglang.srt.layers.moe.ep_moe.utils import (
     moe_kernel_quantize_input)
+from sglang.srt.utils import run_once
+from sglang.srt.distributed.parallel_state import get_world_group
 
+
+from weakref import WeakValueDictionary
+import threading
+
+logger = logging.getLogger(__name__)
+
+
+# Adapted from https://github.com/vllm-project/vllm/blob/f9c069c85e029830094ff9abb926ffbf37b7c7e7/vllm/distributed/parallel_state.py#L940
+PPLX_DID_INIT: bool = False
+
+
+# Adapted from https://github.com/vllm-project/vllm/blob/f9c069c85e029830094ff9abb926ffbf37b7c7e7/vllm/model_executor/layers/fused_moe/layer.py#L277
+class AllToAllCache:
+
+    def __init__(self):
+        self._cache: WeakValueDictionary = WeakValueDictionary()
+        self._lock = threading.RLock()  # Reentrant lock for thread safety
+
+    def destroy(self):
+        with self._lock:
+            # TODO: can we do del self._cache?
+            for _, a2a in self._cache.items():
+                a2a.destroy()
+                pplx_finalize()
+
+    def get_or_create(self, **kwargs):
+
+        # Create a hashable key from the kwargs
+        key = tuple(sorted((k, v) for k, v in kwargs.items()))
+
+        with self._lock:
+            instance = self._cache.get(key)
+            if instance is None:
+                # TODO check if it's right
+                pplx_init(kwargs["rank"], kwargs["world_size"])
+
+                # TODO (varun): Add support to switch to intranode
+                # when all communications are within the same
+                # node.
+                logger.debug("Create AllToAll %s", kwargs)
+                instance = pplx.AllToAll.internode(**kwargs)
+                self._cache[key] = instance
+            return instance
+
+# Global singleton
+_all_to_all_cache = AllToAllCache()
+
+# Factory function as a cleaner interface
+def get_all_to_all(**kwargs):
+    return _all_to_all_cache.get_or_create(**kwargs)
+
+
+
+# Adapted from https://github.com/vllm-project/vllm/blob/f9c069c85e029830094ff9abb926ffbf37b7c7e7/vllm/distributed/parallel_state.py#L944
+@run_once
+def pplx_init(rank, world_size):
+    if world_size > 1:
+        try:
+            global PPLX_DID_INIT
+            logger.debug(
+                "Initialize NVSHMEM for PPLX kernels: rank=%d, "
+                "world size=%d", rank, world_size)
+            uid = nvshmem_get_unique_id() if rank == 0 else nvshmem_alloc_empty_unique_id()
+            uid_gpu = uid.cuda()
+            get_world_group().broadcast(uid_gpu, src=0)
+            uid = uid_gpu.to(device='cpu')
+            logger.debug("PPLX NVSHMEM UID = %s", uid)
+            nvshmem_init(uid, rank, world_size)
+            PPLX_DID_INIT = True
+        except Exception as ex:
+            logger.error("Failed to initialize NVSHMEM for PPLX: %s", ex)
+
+# Adapted from https://github.com/vllm-project/vllm/blob/f9c069c85e029830094ff9abb926ffbf37b7c7e7/vllm/distributed/parallel_state.py#L968
+@run_once
+def pplx_finalize():
+    global PPLX_DID_INIT
+    if PPLX_DID_INIT:
+        logger.debug("PPLX NVSHMEM finalize")
+        _all_to_all_cache.destroy()
+        nvshmem_finalize()
 
 # Note use: layer.get_all_to_all() to get an AllToAll instance
 # The max_num_tokens, world_size and dp_size must be the same
 # as the ones used to create the AllToAll.
-class PplxPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
+class PplxPrepareAndFinalize:
 
     def __init__(self,
                  a2a: pplx.AllToAll,
@@ -21,6 +107,7 @@ class PplxPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
                  world_size: int,
                  rank: int,
                  dp_size: int,
+                 num_experts: int,
                  quant_dtype: Optional[torch.dtype] = None,
                  block_shape: Optional[list[int]] = None):
         super().__init__()
@@ -30,8 +117,15 @@ class PplxPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
         self.max_num_tokens = max_num_tokens
         self.world_size = world_size
         self.rank = rank
-        self.dp_size = dp_size
+        self.num_experts = num_experts
         self.quant_dtype = quant_dtype
+        self.dp_size = dp_size
+
+        # rem_experts need to be 0 for pplx to work properly.
+        rem_experts = self.num_experts % self.world_size
+        assert rem_experts == 0
+        self.num_local_experts = ((self.num_experts // self.world_size) +
+                             (1 if self.rank < rem_experts else 0))
 
     def prepare(
         self,
@@ -40,7 +134,6 @@ class PplxPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
         a2_scale: Optional[torch.Tensor],
         rank_topk_weights: torch.Tensor,
         rank_topk_ids: torch.Tensor,
-        num_experts: int,
         expert_map: Optional[torch.Tensor],
         apply_router_weight_on_input: bool,
     ) -> tuple[torch.Tensor, Optional[torch.Tensor], Optional[torch.Tensor]]:
@@ -68,21 +161,14 @@ class PplxPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
                                                    per_act_token,
                                                    self.block_shape)
 
-        # rem_experts need to be 0 for pplx to work properly.
-        rem_experts = num_experts % self.world_size
-        assert rem_experts == 0
-        num_local_experts = ((num_experts // self.world_size) +
-                             (1 if self.rank < rem_experts else 0))
-
         expert_num_tokens = torch.empty(
-            num_local_experts,
+            self.num_local_experts,
             dtype=torch.int32,
             device=device,
         )
 
-        num_dp = self.world_size // self.dp_size
         expert_x = torch.empty(
-            (num_local_experts, self.max_num_tokens * num_dp, hidden_dim),
+            (self.num_local_experts, self.max_num_tokens * self.dp_size, hidden_dim),
             dtype=a1q.dtype,
             device=device,
         )
@@ -94,7 +180,7 @@ class PplxPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
                           else 1) * float32_size
             expert_x_scale = torch.empty(
                 (
-                    num_experts,
+                    self.num_experts,
                     expert_x.size(1),
                     (expert_x.size(2) + block_size - 1) // block_size,
                 ),
@@ -146,3 +232,45 @@ class PplxPrepareAndFinalize(mk.FusedMoEPrepareAndFinalize):
                          weights=topk_weights,
                          expert_y=fused_expert_output,
                          bound_m=bound_m)
+
+# Adapted from https://github.com/vllm-project/vllm/blob/f9c069c85e029830094ff9abb926ffbf37b7c7e7/vllm/model_executor/layers/fused_moe/layer.py#L683
+def construct_prepare_finalize(
+    max_num_tokens: int,
+    num_experts: int,
+    experts_per_token: int,
+    rank: int,
+    world_size: int,
+    dp_size: int,
+    hidden_dim: int,
+    param_dtype: torch.dtype,
+    block_size: int,
+) -> Optional[PplxPrepareAndFinalize]:
+    logger.debug("using PplxPrepareAndFinalize")
+    assert world_size % dp_size == 0
+
+    all_to_all = get_all_to_all(
+        max_num_tokens=max_num_tokens,
+        num_experts=num_experts,
+        experts_per_token=experts_per_token,  # topk
+        rank=rank,
+        world_size=world_size,
+        dp_size=world_size // dp_size, # dp_size means attn_tp_size in pplx
+        hidden_dim=hidden_dim,
+        hidden_dim_bytes=hidden_dim * param_dtype.itemsize,
+        # For blocked per token: set to
+        #   ceil_div(hidden_dim, block_size) * sizeof(float32)
+        # For per-token: set to sizeof(float32)
+        hidden_dim_scale_bytes=(0 if param_dtype.itemsize != 1 else
+                                ((hidden_dim + block_size - 1) //
+                                 block_size * torch.float32.itemsize)))
+
+    return PplxPrepareAndFinalize(
+        all_to_all,
+        max_num_tokens=max_num_tokens,
+        world_size=world_size,
+        rank=rank,
+        dp_size=dp_size,
+        num_experts=num_experts,
+        quant_dtype=param_dtype,
+    )
+
