+++ python/sglang/srt/managers/schedule_batch.py
@@ -84,6 +84,7 @@ GLOBAL_SERVER_ARGS_KEYS = [
     "enable_dp_lm_head",
     "enable_deepep_moe",
     "deepep_mode",
+    "enable_pplx_moe",
     "enable_ep_moe",
     "moe_dense_tp_size",
     "ep_dispatch_algorithm",
@@ -99,8 +100,6 @@ GLOBAL_SERVER_ARGS_KEYS = [
     "torchao_config",
     "triton_attention_reduce_in_fp32",
     "num_reserved_decode_tokens",
-    "max_tokens_across_dp",
-    "enable_pplx_moe",
 ]
 
 # Put some global args for easy access
